requirements.txt
numpy==1.23.3
scikit-image
opencv-python
tensorboard
torch
torchvision
configparser
matplotlib
Pillow